dark mode and audio buffering actually works

+++ lib/providers/song_list_provider.dart
@@ -26,7 +26,7 @@ class Song {
   String songName;
   String artist;
   int duration;
-  int price;
+  BigInt price;
   int byteSize;
   Uint8List songId;
   DistributorContact? distributorContact;
